Decompose recovery_engine manual-rebind path (#3834)
Behavior-preserving decomposition of the giant
`src/services/discord/recovery_engine.rs`. Extracts the manual-rebind recovery
path — `rebind_inflight_for_channel` and its private support cluster
(`PendingCodexTuiRebindRelay` / `PendingRebindInflightRollback` carriers + the
`codex_tui_*` resume-offset / replay-event helpers) and the path's unit tests —
verbatim into the new leaf submodule `recovery_engine/manual_rebind.rs`.

Extraction map (fn cluster -> new module):
- recovery_engine/manual_rebind.rs (NEW, 889 prod / 836 test LoC):
  rebind_inflight_for_channel, codex_tui_* (9 helpers), Pending* (2 types),
  post_work_evidence_tests.
- recovery_engine.rs: 3899 -> 3039 prod LoC (raw 5363 -> 3667).

Behavior-preserving invariant: pure code movement + visibility plumbing only.
`rebind_inflight_for_channel` is re-exported (`pub(crate) use`) so the external
path `recovery_engine::rebind_inflight_for_channel` stays byte-identical for its
`health` caller. `RebindOutcome` / `RebindError` are intentionally KEPT in the
root module because the sibling `recovery_engine/rebind_runtime.rs` and external
callers (`health`, `health::recovery`) reference them. The moved code uses
`use super::*`, mirroring the eight existing leaf submodules; all `super::`
paths re-resolve through the root's glob re-export, so no path rewrites were
needed. No logic, signature, or behavior change.

Gates: cargo fmt --check clean; cargo check --lib exit 0 (no new warnings);
cargo test --lib recovery_engine 75 passed; agent-maintenance freshness check
passed; audit_maintainability --check exit 0 (giant baseline synced
3899 -> 3039); check_hotfile_ratchet exit 0 (recovery_engine.rs is not a #3016
hotfile and is unaffected).
